Criando player

+++ dr-map/src/lib.rs
@@ -1,6 +1,6 @@
-const SCREEN_WIDTH: usize = 80;
-const SCREEN_HEIGHT: usize = 50;
-const NUM_TILES: usize = SCREEN_WIDTH * SCREEN_HEIGHT;
+const SCREEN_WIDTH: i32 = 80;
+const SCREEN_HEIGHT: i32 = 50;
+const NUM_TILES: usize = (SCREEN_WIDTH * SCREEN_HEIGHT) as usize;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum TileType {
